chore: clear all ESLint errors for issue #185 lint debt zero
- eslint.config.js: ignore coverage/, add varsIgnorePattern/argsIgnorePattern/destructuredArrayIgnorePattern '^_'
- HexGrid.tsx: extract findHexAtClientXY to hexGridUtils.ts (react-refresh), add exhaustive deps (editable, onEditCellClick) to fix react-compiler, inline-disable set-state-in-effect x2 with rationale
- App.tsx: inline-disable react-hooks/refs on usePrevious (intentional stale-ref pattern)
- useModal.ts: inline-disable react-hooks/refs on stable-ref pattern
- codec.test.ts: replace any with Record<string,unknown> / unknown as string
- devExpose.ts: replace window as any with typed window extension
- e2e specs: fix unused vars (_prefix), use GamePhase enum instead of 'DrawCard' as any, inline-disable window as any in page.evaluate browser context
- ci.yml: remove continue-on-error: true from Lint step

Result: 0 errors (12 warnings remain, not in scope), 786 unit tests pass, 56 e2e pass / 9 skip / 0 fail, qa:smoke 7/7
