+++ .github/workflows/ci.yml
@@ -28,7 +28,6 @@ jobs:
 
       - name: Lint
         run: npm run lint
-        continue-on-error: true
 
       - name: Run tests with coverage
         run: npm run test:coverage
