+++ src/components/Board/HexGrid.tsx
@@ -1,73 +1,14 @@
 import React, { useState, useRef, useCallback, useEffect } from 'react';
 import { Board } from '../../core/board';
-import { AxialCoord, hexEquals, HEX_SIZE, axialToPixel } from '../../core/hex';
+import { AxialCoord, hexEquals, HEX_SIZE } from '../../core/hex';
 import { HexCell } from './HexCell';
 import { TerrainDefs } from './TerrainDefs';
 import { PieceDefs } from './PieceDefs';
-import { Player, HexCell as HexCellData } from '../../types';
-import { useBoardTransform, Transform } from '../../hooks/useBoardTransform';
+import { Player } from '../../types';
+import { useBoardTransform } from '../../hooks/useBoardTransform';
 import { useTranslation } from 'react-i18next';
 import { useGameStore } from '../../store/gameStore';
-
-/**
- * Given a client-space coordinate (e.g. from mouse or touch event),
- * reverse the container transform and SVG viewBox scaling to find
- * which hex cell is at that point. Returns null if no cell is close enough.
- */
-export function findHexAtClientXY(
-  clientX: number,
-  clientY: number,
-  cells: HexCellData[],
-  transform: Transform,
-  containerRect: DOMRect,
-  svgElement: SVGSVGElement,
-  gridOffset: number,
-): AxialCoord | null {
-  // Step 1: position relative to container
-  const cx = clientX - containerRect.left;
-  const cy = clientY - containerRect.top;
-
-  // Step 2: undo transform (scale + translate)
-  const sx = (cx - transform.translateX) / transform.scale;
-  const sy = (cy - transform.translateY) / transform.scale;
-
-  // Step 3: SVG viewBox scaling
-  // The inner <div> is 100%×100% of container, SVG is also 100%×100% of that div
-  // We need the ratio between the rendered SVG size and its viewBox
-  const svgRect = svgElement.getBoundingClientRect();
-  const viewBox = svgElement.viewBox.baseVal;
-  const scaleX = svgRect.width > 0 ? viewBox.width / svgRect.width : 1;
-  const scaleY = svgRect.height > 0 ? viewBox.height / svgRect.height : 1;
-
-  // sx/sy are in the coordinate space of the inner div (which is 100%x100% of container)
-  // svgRect is also in client space, so we need to account for svgRect offset relative to container
-  const svgOffsetX = svgRect.left - containerRect.left;
-  const svgOffsetY = svgRect.top - containerRect.top;
-
-  const viewBoxX = (sx - svgOffsetX) * scaleX;
-  const viewBoxY = (sy - svgOffsetY) * scaleY;
-
-  // Step 4: subtract gridOffset to get hex coordinate space
-  const hexSpaceX = viewBoxX - gridOffset;
-  const hexSpaceY = viewBoxY - gridOffset;
-
-  // Step 5: find closest cell within HEX_SIZE * 0.95 threshold
-  let bestCell: HexCellData | null = null;
-  let bestDist = HEX_SIZE * 0.95;
-
-  for (const cell of cells) {
-    const center = axialToPixel(cell.coord, HEX_SIZE);
-    const dx = hexSpaceX - center.x;
-    const dy = hexSpaceY - center.y;
-    const dist = Math.sqrt(dx * dx + dy * dy);
-    if (dist < bestDist) {
-      bestDist = dist;
-      bestCell = cell;
-    }
-  }
-
-  return bestCell ? bestCell.coord : null;
-}
+import { findHexAtClientXY } from './hexGridUtils';
 
 interface HexGridProps {
   board: Board;
@@ -138,6 +79,7 @@ export const HexGrid: React.FC<HexGridProps> = React.memo(({
 
   useEffect(() => {
     if (!invalidClickKey) return;
+    // eslint-disable-next-line react-hooks/set-state-in-effect -- UI-only: sync invalidClickKey to local flash state for one-shot animation
     setInvalidFlashKey(invalidClickKey);
     const timer = setTimeout(() => setInvalidFlashKey(null), 400);
     return () => clearTimeout(timer);
@@ -171,6 +113,7 @@ export const HexGrid: React.FC<HexGridProps> = React.memo(({
     prevSettlementKeysRef.current = currentKeys;
 
     if (newKey) {
+      // eslint-disable-next-line react-hooks/set-state-in-effect -- UI-only: sync newly-placed settlement key to local state for one-shot ring animation
       setRecentlyPlacedKey(newKey);
       const timer = setTimeout(() => setRecentlyPlacedKey(null), 350);
       return () => clearTimeout(timer);
@@ -260,7 +203,7 @@ export const HexGrid: React.FC<HexGridProps> = React.memo(({
         }
       }
     },
-    [board, validPlacements, onCellClick, onEscape, focusHex],
+    [board, validPlacements, onCellClick, onEscape, focusHex, editable, onEditCellClick],
   );
 
   // Determine the single entry-point cell that gets tabIndex=0 (roving tabindex pattern)
