+++ src/mapEditor/devExpose.ts
@@ -2,5 +2,9 @@ import { encode, decode } from './codec';
 import { useCustomMapStore } from './customMapStore';
 
 if (import.meta.env.DEV) {
-  (window as any).__KB__ = { encode, decode, store: () => useCustomMapStore.getState() };
+  (window as typeof window & { __KB__: unknown }).__KB__ = {
+    encode,
+    decode,
+    store: () => useCustomMapStore.getState(),
+  };
 }
